requirements.txt
pydantic
fastapi
uvicorn
python-docx
httpx
python-dotenv
tiktoken
numpy
scikit-learn
neo4j
qdrant-client
